+++ docs/ReleaseNotes.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - Minor fixes and code formatting improvements.
 - Added information message.
+- Improve cleaning function.
 
 ## [3.6.1] - 2024.07.25